fixed secret to keyvault & apiversion for RTP & delete deployment from 500 to 409 & some updated comment and error message

+++ end-to-end-solutions/Luna/src/Luna.API/Controllers/Admin/Luna.AI/DeploymentController.cs
@@ -25,16 +25,18 @@ namespace Luna.API.Controllers.Admin
     {
         private readonly IDeploymentService _deploymentService;
         private readonly ILogger<RestrictedUserController> _logger;
+        private readonly IAPIVersionService _apiVersionService;
 
         /// <summary>
         /// Constructor that uses dependency injection.
         /// </summary>
         /// <param name="deploymentService">The service to inject.</param>
         /// <param name="logger">The logger.</param>
-        public DeploymentController(IDeploymentService deploymentService, ILogger<RestrictedUserController> logger)
+        public DeploymentController(IDeploymentService deploymentService, ILogger<RestrictedUserController> logger, IAPIVersionService apiVersionService)
         {
             _deploymentService = deploymentService ?? throw new ArgumentNullException(nameof(deploymentService));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _apiVersionService = apiVersionService ?? throw new ArgumentNullException(nameof(apiVersionService));
         }
 
         /// <summary>
@@ -116,6 +118,14 @@ namespace Luna.API.Controllers.Admin
         {
             AADAuthHelper.VerifyUserAccess(this.HttpContext, _logger, true);
             _logger.LogInformation($"Delete deployment {deploymentName} from product {productName}.");
+
+            // check if there exist apiversions
+            var apiVersions = await _apiVersionService.GetAllAsync(productName, deploymentName);
+            if (apiVersions.Count != 0)
+            {
+                throw new LunaConflictUserException($"Unable to delete {deploymentName} with subscription");
+            }
+
             await _deploymentService.DeleteAsync(productName, deploymentName);
             return NoContent();
         }
